requirements.txt

pillow
pandas
opencv-python
scikit-image
matplotlib
django==4.1.7
tensorflow